+++ src/renderer/markup/ast-converter/entities/class.ts
@@ -1,7 +1,7 @@
 import { convertSeeTagsForDocumentation } from "unwritten:renderer/markup/ast-converter/shared/see";
 import { registerAnchor, registerAnonymousAnchor } from "unwritten:renderer/markup/registry/registry";
 import { renderMemberContext, withMemberContext } from "unwritten:renderer/markup/utils/context";
-import { renderEntityPrefix } from "unwritten:renderer/markup/utils/renderer.js";
+import { isRenderObjectMemberTitlesEnabled, renderEntityPrefix } from "unwritten:renderer/markup/utils/renderer.js";
 import { getRenderConfig } from "unwritten:renderer/utils/config";
 import { filterOutImplicitSignatures, filterOutPrivateMembers } from "unwritten:renderer/utils/private-members";
 import {
@@ -81,48 +81,60 @@ export function convertClassEntityForTableOfContents(ctx: MarkupRenderContexts, 
 
     const constructorEntity = extendClassEntityConstructorsWithHeritage(classEntity);
     const propertyEntities = extendClassEntityEntitiesWithHeritage(classEntity, "properties");
-    const eventPropertyEntities = extendClassEntityEntitiesWithHeritage(classEntity, "events");
     const methodEntities = extendClassEntityEntitiesWithHeritage(classEntity, "methods");
     const setterEntities = extendClassEntityEntitiesWithHeritage(classEntity, "setters");
     const getterEntities = extendClassEntityEntitiesWithHeritage(classEntity, "getters");
+    const eventPropertyEntities = extendClassEntityEntitiesWithHeritage(classEntity, "events");
 
     const publicConstructorEntity = renderConfig.renderPrivateMembers ? constructorEntity : constructorEntity && filterOutPrivateMembers([constructorEntity])[0];
     const publicPropertyEntities = renderConfig.renderPrivateMembers ? propertyEntities : filterOutPrivateMembers(propertyEntities);
-    const publicEventPropertyEntities = renderConfig.renderPrivateMembers ? eventPropertyEntities : filterOutPrivateMembers(eventPropertyEntities);
     const publicMethodEntities = renderConfig.renderPrivateMembers ? methodEntities : filterOutPrivateMembers(methodEntities);
     const publicSetterEntities = renderConfig.renderPrivateMembers ? setterEntities : filterOutPrivateMembers(setterEntities);
     const publicGetterEntities = renderConfig.renderPrivateMembers ? getterEntities : filterOutPrivateMembers(getterEntities);
+    const publicEventPropertyEntities = renderConfig.renderPrivateMembers ? eventPropertyEntities : filterOutPrivateMembers(eventPropertyEntities);
 
     const explicitConstructSignatures = publicConstructorEntity?.signatures && filterOutImplicitSignatures(publicConstructorEntity.signatures);
     const convertedConstructSignatures = explicitConstructSignatures?.map(signatureEntity => convertSignatureEntityForTableOfContents(ctx, signatureEntity));
     const convertedProperties = publicPropertyEntities.map(propertyEntity => convertPropertyEntityForTableOfContents(ctx, propertyEntity));
-    const convertedEventProperties = publicEventPropertyEntities.map(eventPropertyEntity => convertEventPropertyEntityForTableOfContents(ctx, eventPropertyEntity));
     const convertedMethods = publicMethodEntities.flatMap(methodEntity => convertFunctionLikeEntityForTableOfContents(ctx, methodEntity)).flat();
     const convertedSetters = publicSetterEntities.flatMap(setterEntity => convertFunctionLikeEntityForTableOfContents(ctx, setterEntity));
     const convertedGetters = publicGetterEntities.flatMap(getterEntity => convertFunctionLikeEntityForTableOfContents(ctx, getterEntity));
+    const convertedEventProperties = publicEventPropertyEntities.map(eventPropertyEntity => convertEventPropertyEntityForTableOfContents(ctx, eventPropertyEntity));
 
     const constructorList = convertedConstructSignatures && convertedConstructSignatures.length > 0
-      ? [translate("ctor", { capitalize: true, count: convertedConstructSignatures.length }), createListNode(...convertedConstructSignatures)]
+      ? isRenderObjectMemberTitlesEnabled(ctx, "tableOfContents")
+        ? [translate("ctor", { capitalize: true, count: convertedConstructSignatures.length }), createListNode(...convertedConstructSignatures)]
+        : [createListNode(...convertedConstructSignatures)]
       : [];
 
     const propertyList = convertedProperties.length > 0
-      ? [translate("property", { capitalize: true, count: convertedProperties.length }), createListNode(...convertedProperties)]
-      : [];
-
-    const eventPropertyList = convertedEventProperties.length > 0
-      ? [translate("event", { capitalize: true, count: convertedEventProperties.length }), createListNode(...convertedEventProperties)]
+      ? isRenderObjectMemberTitlesEnabled(ctx, "tableOfContents")
+        ? [translate("property", { capitalize: true, count: convertedProperties.length }), createListNode(...convertedProperties)]
+        : [createListNode(...convertedProperties)]
       : [];
 
     const methodList = convertedMethods.length > 0
-      ? [translate("method", { capitalize: true, count: convertedMethods.length }), createListNode(...convertedMethods)]
+      ? isRenderObjectMemberTitlesEnabled(ctx, "tableOfContents")
+        ? [translate("method", { capitalize: true, count: convertedMethods.length }), createListNode(...convertedMethods)]
+        : [createListNode(...convertedMethods)]
       : [];
 
     const setterList = convertedSetters.length > 0
-      ? [translate("setter", { capitalize: true, count: convertedSetters.length }), createListNode(...convertedSetters)]
+      ? isRenderObjectMemberTitlesEnabled(ctx, "tableOfContents")
+        ? [translate("setter", { capitalize: true, count: convertedSetters.length }), createListNode(...convertedSetters)]
+        : [createListNode(...convertedSetters)]
       : [];
 
     const getterList = convertedGetters.length > 0
-      ? [translate("getter", { capitalize: true, count: convertedGetters.length }), createListNode(...convertedGetters)]
+      ? isRenderObjectMemberTitlesEnabled(ctx, "tableOfContents")
+        ? [translate("getter", { capitalize: true, count: convertedGetters.length }), createListNode(...convertedGetters)]
+        : [createListNode(...convertedGetters)]
+      : [];
+
+    const eventPropertyList = convertedEventProperties.length > 0
+      ? isRenderObjectMemberTitlesEnabled(ctx, "tableOfContents")
+        ? [translate("event", { capitalize: true, count: convertedEventProperties.length }), createListNode(...convertedEventProperties)]
+        : [createListNode(...convertedEventProperties)]
       : [];
 
     return [
@@ -188,41 +200,48 @@ export function convertClassEntityForDocumentation(ctx: MarkupRenderContexts, cl
 
     const hasConstructSignatures = explicitConstructSignatures && explicitConstructSignatures.length > 0;
     const constructSignaturesTranslation = translate("constructSignature", { capitalizeEach: true, count: explicitConstructSignatures?.length });
-    const constructSignaturesAnchor = hasConstructSignatures && registerAnonymousAnchor(ctx, constructSignaturesTranslation);
+    const constructSignaturesAnchor = isRenderObjectMemberTitlesEnabled(ctx, "documentation") && hasConstructSignatures && registerAnonymousAnchor(ctx, constructSignaturesTranslation);
 
     const hasProperties = publicPropertyEntities.length > 0;
     const propertiesTranslation = translate("property", { capitalizeEach: true, count: publicPropertyEntities.length });
-    const propertiesAnchor = hasProperties && registerAnonymousAnchor(ctx, propertiesTranslation);
+    const propertiesAnchor = isRenderObjectMemberTitlesEnabled(ctx, "documentation") && hasProperties && registerAnonymousAnchor(ctx, propertiesTranslation);
 
     const hasMethods = publicMethodEntities.length > 0;
     const methodsTranslation = translate("method", { capitalizeEach: true, count: publicMethodEntities.length });
-    const methodsAnchor = hasMethods && registerAnonymousAnchor(ctx, methodsTranslation);
+    const methodsAnchor = isRenderObjectMemberTitlesEnabled(ctx, "documentation") && hasMethods && registerAnonymousAnchor(ctx, methodsTranslation);
 
     const hasSetters = publicSetterEntities.length > 0;
     const settersTranslation = translate("setter", { capitalizeEach: true, count: publicSetterEntities.length });
-    const settersAnchor = hasSetters && registerAnonymousAnchor(ctx, settersTranslation);
+    const settersAnchor = isRenderObjectMemberTitlesEnabled(ctx, "documentation") && hasSetters && registerAnonymousAnchor(ctx, settersTranslation);
 
     const hasGetters = publicGetterEntities.length > 0;
     const gettersTranslation = translate("getter", { capitalizeEach: true, count: publicGetterEntities.length });
-    const gettersAnchor = hasGetters && registerAnonymousAnchor(ctx, gettersTranslation);
+    const gettersAnchor = isRenderObjectMemberTitlesEnabled(ctx, "documentation") && hasGetters && registerAnonymousAnchor(ctx, gettersTranslation);
 
     const hasEventProperties = publicEventPropertyEntities.length > 0;
     const eventPropertiesTranslation = translate("event", { capitalizeEach: true, count: publicEventPropertyEntities.length });
-    const eventPropertiesAnchor = hasEventProperties && registerAnonymousAnchor(ctx, eventPropertiesTranslation);
+    const eventPropertiesAnchor = isRenderObjectMemberTitlesEnabled(ctx, "documentation") && hasEventProperties && registerAnonymousAnchor(ctx, eventPropertiesTranslation);
 
     const convertedConstructSignatures = explicitConstructSignatures?.map(signatureEntity => convertSignatureEntityForDocumentation(ctx, signatureEntity)) ?? [];
     const convertedProperties = publicPropertyEntities.map(propertyEntity => convertPropertyEntityForDocumentation(ctx, propertyEntity));
-    const convertedEventProperties = publicEventPropertyEntities.map(eventPropertyEntity => convertEventPropertyEntityForDocumentation(ctx, eventPropertyEntity));
     const convertedMethods = publicMethodEntities.flatMap(methodEntity => convertFunctionLikeEntityForDocumentation(ctx, methodEntity));
     const convertedSetters = publicSetterEntities.flatMap(setterEntity => convertFunctionLikeEntityForDocumentation(ctx, setterEntity));
     const convertedGetters = publicGetterEntities.flatMap(getterEntity => convertFunctionLikeEntityForDocumentation(ctx, getterEntity));
+    const convertedEventProperties = publicEventPropertyEntities.map(eventPropertyEntity => convertEventPropertyEntityForDocumentation(ctx, eventPropertyEntity));
 
-    const constructSignatures = constructSignaturesAnchor && createSectionNode("constructor", createTitleNode(constructSignaturesTranslation, constructSignaturesAnchor, ...convertedConstructSignatures));
-    const properties = propertiesAnchor && createSectionNode("properties", createTitleNode(propertiesTranslation, propertiesAnchor, ...convertedProperties));
-    const methods = methodsAnchor && createSectionNode("methods", createTitleNode(methodsTranslation, methodsAnchor, ...convertedMethods));
-    const setters = settersAnchor && createSectionNode("setters", createTitleNode(settersTranslation, settersAnchor, ...convertedSetters));
-    const getters = gettersAnchor && createSectionNode("getters", createTitleNode(gettersTranslation, gettersAnchor, ...convertedGetters));
-    const eventProperties = eventPropertiesAnchor && createSectionNode("events", createTitleNode(eventPropertiesTranslation, eventPropertiesAnchor, ...convertedEventProperties));
+    const constructSignaturesTitle = constructSignaturesAnchor && createTitleNode(constructSignaturesTranslation, constructSignaturesAnchor, ...convertedConstructSignatures);
+    const propertiesTitle = propertiesAnchor && createTitleNode(propertiesTranslation, propertiesAnchor, ...convertedProperties);
+    const methodsTitle = methodsAnchor && createTitleNode(methodsTranslation, methodsAnchor, ...convertedMethods);
+    const settersTitle = settersAnchor && createTitleNode(settersTranslation, settersAnchor, ...convertedSetters);
+    const gettersTitle = gettersAnchor && createTitleNode(gettersTranslation, gettersAnchor, ...convertedGetters);
+    const eventPropertiesTitle = eventPropertiesAnchor && createTitleNode(eventPropertiesTranslation, eventPropertiesAnchor, ...convertedEventProperties);
+
+    const constructSignatures = hasConstructSignatures && createSectionNode("constructor", ...constructSignaturesTitle ? [constructSignaturesTitle] as const : convertedConstructSignatures);
+    const properties = hasProperties && createSectionNode("properties", ...propertiesTitle ? [propertiesTitle] as const : convertedProperties);
+    const methods = hasMethods && createSectionNode("methods", ...methodsTitle ? [methodsTitle] as const : convertedMethods);
+    const setters = hasSetters && createSectionNode("setters", ...settersTitle ? [settersTitle] as const : convertedSetters);
+    const getters = hasGetters && createSectionNode("getters", ...gettersTitle ? [gettersTitle] as const : convertedGetters);
+    const eventProperties = hasEventProperties && createSectionNode("events", ...eventPropertiesTitle ? [eventPropertiesTitle] as const : convertedEventProperties);
 
     return createSectionNode(
       getSectionType(classEntity.kind),